+++ CHANGELOG.md
@@ -48,6 +48,14 @@
         -   The following events have been removed:
             -   `onConnected()`
             -   `onDisconnected()`
+    -   Added in tags to change the state of the inventory's camera controls:
+        -   `aux.context.inventory.pannable` enables and disables the inventory's ability to pan, off by default.
+        -   `aux.context.inventory.resizable` enables and disables the inventory's drag to resize functionality, on by default.
+        -   `aux.context.inventory.rotatable` enables and disables the inventory's ability to rotate, on by default.
+        -   `aux.context.inventory.zoomable` enables and disables the inventory's ability to zoom, on by default.
+-   Bug Fixes
+    -   Resolved issue with the near cliiping plane for the sheet's minifile image.
+    -   Resolved issues with the create empty bot button not functioning sometimes on mobile.
 
 ## V0.9.40